FIX - Tidy Machine Readable Report - moved ItemDetails to tidy namespacing; - removed namespaces from FeatureReport; and - made XSLT TransformerFactory static.
